# perf(links): improve drastically perfs for update links inside table-mode