--- a/antarest/study/business/link_management.py
+++ b/antarest/study/business/link_management.py
@@ -70,7 +70,9 @@ class LinkManager:
 
         return link_creation_dto
 
-    def update_link(self, study: StudyInterface, area_from: str, area_to: str, link_update_dto: LinkBaseDTO) -> LinkDTO:
+    def _create_update_link_command(
+        self, study: StudyInterface, area_from: str, area_to: str, link_update_dto: LinkBaseDTO
+    ) -> tuple[UpdateLink, LinkInternal]:
         link_dto = LinkDTO(area1=area_from, area2=area_to, **link_update_dto.model_dump(exclude_unset=True))
 
         file_study = study.get_files()
@@ -87,6 +89,10 @@ class LinkManager:
             command_context=self._command_context,
             study_version=study.version,
         )
+        return command, link
+
+    def update_link(self, study: StudyInterface, area_from: str, area_to: str, link_update_dto: LinkBaseDTO) -> LinkDTO:
+        command, link = self._create_update_link_command(study, area_from, area_to, link_update_dto)
 
         study.add_commands([command])
 
@@ -102,22 +108,7 @@ class LinkManager:
         # Build all commands
         commands = []
         for (area1, area2), update_link_dto in update_links_by_ids.items():
-            link_dto = LinkDTO(area1=area1, area2=area2, **update_link_dto.model_dump(exclude_unset=True))
-
-            file_study = study.get_files()
-            link = link_dto.to_internal(study.version)
-
-            self._get_link_if_exists(file_study, link)
-
-            command = UpdateLink(
-                area1=link.area1,
-                area2=link.area2,
-                parameters=link.model_dump(
-                    include=update_link_dto.model_fields_set, exclude={"area1", "area2"}, exclude_none=True
-                ),
-                command_context=self._command_context,
-                study_version=study.version,
-            )
+            command = self._create_update_link_command(study, area1, area2, update_link_dto)[0]
             commands.append(command)
 
         study.add_commands(commands)