Benchmark:

Big BP study (330 links):
- RAW:
    - Before:  7,14 s
    - After: 4,39 s
- VARIANT:
    - Before: 326 s
    - After: 19,6 s